requirements.txt
streamlit==1.38.0  # pinned for static asset stability on Render
pdf2image
Pillow
python-dotenv
groq
google-generativeai
streamlit-option-menu
plotly
# poppler-utils is a system package; keeping line for reference but not installable via pip.
# Remove or handle externally in deployment. pdf2image may require it; on Render consider disabling preview features.
poppler-utils
crewai
opencv-python-headless
openai>=1.40.0
streamlit-extras
pytest
requests
vosk
pyttsx3
streamlit-audiorecorder
pydub
gTTS 
watchdog
langchain-openai
langchain-groq
langchain-google-genai
edge-tts  # New primary TTS provider (Azure Edge voices)
mistune
beautifulsoup4